Add Windows x86_64 nightly release path
Merge the reviewed unsigned Windows release path and protected publisher preflight gate.

+++ src/vst3/transport.rs
@@ -5,6 +5,7 @@ use toybox::vst3::prelude::*;
 use crate::capture::TransportInfo;
 
 /// Convert one VST3 process context without inventing missing host facts.
+#[cfg_attr(not(target_os = "windows"), allow(clippy::unnecessary_cast))]
 pub(super) unsafe fn transport_from_process_context(
     process_context: *mut ProcessContext,
 ) -> TransportInfo {
@@ -12,11 +13,12 @@ pub(super) unsafe fn transport_from_process_context(
         return TransportInfo::default();
     };
     let state = context.state;
-    let tempo_valid = (state & ProcessContext_::StatesAndFlags_::kTempoValid) != 0;
-    let position_valid = (state & ProcessContext_::StatesAndFlags_::kProjectTimeMusicValid) != 0;
+    let tempo_valid = (state & (ProcessContext_::StatesAndFlags_::kTempoValid as u32)) != 0;
+    let position_valid =
+        (state & (ProcessContext_::StatesAndFlags_::kProjectTimeMusicValid as u32)) != 0;
     TransportInfo {
         tempo_bpm: tempo_valid.then_some(context.tempo),
         song_pos_beats: position_valid.then_some(context.projectTimeMusic),
-        is_playing: (state & ProcessContext_::StatesAndFlags_::kPlaying) != 0,
+        is_playing: (state & (ProcessContext_::StatesAndFlags_::kPlaying as u32)) != 0,
     }
 }
